chore(tests): move ParseIntoTableApiPlugin to plugin_collection
