+++ falco_plugin_tests/src/plugin_collection/parse/mod.rs
@@ -1,1 +1,2 @@
+pub mod remaining_into_table_api;
 pub mod remaining_into_table_direct;
